+++ include/movable.hh
@@ -11,11 +11,11 @@ namespace abstract
     {
         public :
             Movable() = default;
-            virtual void impulse(sf::Vector2f intensity) = 0;
+            virtual void impulse(sf::Vector2i intensity) = 0;
             virtual void friction(void) = 0;
-            sf::Vector2f velocity_get(void) { return velocity_; }
+            sf::Vector2i velocity_get(void) { return velocity_; }
 
         protected :
-            sf::Vector2f velocity_ = sf::Vector2f(0, 0);
+            sf::Vector2i velocity_ = sf::Vector2i(0, 0);
     };
 }
