+++ src/obstacle.cc
@@ -21,7 +21,7 @@ namespace level
     {}
 
     Obstacle::Obstacle(std::string path_to_texture,
-            sf::Vector2f pos,
+            sf::Vector2i pos,
             sf::Vector2i len,
             sf::Vector2i sprite_origin
             )
