+++ tests/object/object_tests.hh
@@ -8,17 +8,17 @@
 class CtrObjectTesterBadCoordinates : public testing::TestWithParam<int*>
 {};
 
-class CtrObjectTesterPairBadCoordinates : public testing::TestWithParam<sf::Vector2f*>
+class CtrObjectTesterPairBadCoordinates : public testing::TestWithParam<sf::Vector2i*>
 {};
 
 class CtrObjectTesterGoodCoordinates : public testing::TestWithParam<int*>
 {};
 
-class CtrObjectTesterPairGoodCoordinates : public testing::TestWithParam<sf::Vector2f*>
+class CtrObjectTesterPairGoodCoordinates : public testing::TestWithParam<sf::Vector2i*>
 {};
 
 class CtrObjectTesterBadLen : public testing::TestWithParam<int*>
 {};
 
-class CtrObjectTesterPairBadLen : public testing::TestWithParam<sf::Vector2f*>
+class CtrObjectTesterPairBadLen : public testing::TestWithParam<sf::Vector2i*>
 {};
